Add CODEGEN keyword to add_custom_command() if supported
If using CMake 3.31 or later, add_custom_command() supports a CODEGEN
keyword. It requires policy CMP0171 to be set to NEW to be recognized.
Add the CODEGEN keyword to these calls if we can so that consumers can
build the `codegen` CMake target to ensure code generation has been
performed without having to actually compile the generated code. This is
useful for linting CI jobs like clang-tidy.